Improved code readability and added more comments
refactored try catch block to make it a simple if-else statement where the if block handles the successful handling and the else block handles all errors.

Now, all we only serialise the data if and when needed.

Updated comments for the corresponding sections as well so that it explains code base in a much more detailed way.

TODO/FIX:
- Implement feature to allow users to input a post code of their choice

- Add more comments and help string

- Make further efficiencies where possible